% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMathInput.R
\name{updateMathInput}
\alias{updateMathInput}
\title{Chabnge the value of a mathematical input on the client}
\usage{
updateMathInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}
}
\description{
Chabnge the value of a mathematical input on the client
}
